refactor: mdx and src used components
