--- apps/blog/src/app/components/pages/[post]/components/mdx-custom/link/GlowingLink.tsx
@@ -1,17 +1,0 @@
-import { GlowingText } from "@ashgw/components";
-
-import { BlogLink } from "./Link";
-
-export const GlowingBlogLink = ({
-  href,
-  name,
-}: {
-  href: string;
-  name: string;
-}) => {
-  return (
-    <BlogLink href={href}>
-      <GlowingText>{name}</GlowingText>
-    </BlogLink>
-  );
-};
