+++ apps/blog/src/app/components/pages/[post]/components/mdx-custom/link/index.tsx
@@ -1,2 +1,1 @@
 export { BlogLink } from "./Link";
-export { GlowingBlogLink } from "./GlowingLink";
