+++ apps/blog/src/app/components/pages/[post]/components/mdx-custom/text/TextContent.tsx
