+++ apps/blog/src/app/components/pages/[post]/components/mdx-custom/text/index.tsx
@@ -1,2 +1,1 @@
-export { GlowingText } from "./GlowingText";
 export { TextContent } from "./TextContent";
