+++ apps/www/src/app/components/pages/about/components/Link.tsx
@@ -2,17 +2,18 @@
 
 import NextLink from "next/link";
 
-import { GlowingText } from "@ashgw/components";
-
 interface LinkProps {
   readonly href: string;
   readonly name: string;
   readonly inNewTab?: boolean;
 }
+
 const Link = ({ href, name, inNewTab = false }: LinkProps) => {
   return (
     <NextLink href={href} target={inNewTab ? "_blank" : undefined}>
-      <GlowingText>{name}</GlowingText>
+      <span> </span>
+      <strong className="glows text-white">{name}</strong>
+      <span> </span>
     </NextLink>
   );
 };
