+++ packages/components/src/reusables/index.ts
@@ -1,3 +1,2 @@
-export * from "./text";
 export { LoadingScreen } from "./LoadingScreen";
 export * from "./cookies";
